xml/parsing/manakai/doctypes-1.dat

#data
<!DOCTYPE abcd><abcd/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE abcd>
| <{} abcd>

#data
<!DOCTYPE Abcd><Abcd/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE Abcd>
| <{} Abcd>

#data
<!DOCTYPE ABCD><ABCD/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE ABCD>
| <{} ABCD>

#data
<!DOCTYPE HtML>
<HtML/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE HtML>
| <{} HtML>

#data
<!DOCTYPE html>
<HTML/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE html>
| <{} HTML>

#data
<!doctype html><html/>
#errors
1;1;no XML decl;;;s
1;3;lowercase keyword;DOCTYPE;;m
#document
| <!DOCTYPE html>
| <{} html>

#data
<!DOCTYPe html><html/>
#errors
1;1;no XML decl;;;s
1;3;lowercase keyword;DOCTYPE;;m
#document
| <!DOCTYPE html>
| <{} html>
